pyproject.toml
[tool.poetry]
name = "touric-rest-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-extensions = "^3.2.3"
django-ninja = "^1.1.0"
python-decouple = "^3.8"
django-environ = "^0.11.2"
django-split-settings = "^1.3.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
